This commit was manufactured by cvs2svn to create tag 'PPR-1_44'.
Sprout from vendor 2001-06-18 18:22:22 UTC chappell 'Import of changes made to old source directory while still trying to get'
Cherrypick from master 2001-07-31 20:47:21 UTC chappell '*** empty log message ***':
    .restore_symlinks
    Configure
    Makefile
    docs/Makefile
    docs/misc/History.txt
    docs/misc/announce.txt
    docs/misc/docs_history.txt
    docs/misc/ppr.lsm
    docs/misc/release_steps.html
    docs/pprdoc/Makefile
    docs/refman/Makefile
    docs/refman/papsrv.8.pod
    docs/refman/ppad.8.pod
    docs/rules.mk
    filter_dotmatrix/.depend
    filter_lp/filter_lp.c
    fixup/fixup_inetd.sh
    fixup/fixup_obsolete.sh
    include/gu.h
    include/libppr_int.h
    include/queueinfo.h
    include/version.h
    interfaces/.depend
    interfaces/README.txt
    interfaces/parallel.c
    interfaces/parallel_linux.c
    interfaces/serial.c
    libgu/.depend
    libgu/getline.c
    libgu/ini_section.c
    libgu/malloc.c
    libppr/.depend
    libppr/Makefile
    libppr/int_snmp.c
    libppr/options.c
    libppr/query.c
    libppr/queueinfo.c
    libpprdb/.depend
    libscript/.depend
    libscript/getservbyname.c
    libscript/time_elapsed.c
    libttf/.depend
    libuprint/.depend
    lprsrv/.depend
    makeprogs/.depend
    makeprogs/Makefile
    makeprogs/gen_paths.c
    makeprogs/global.mk
    makeprogs/global.mk.unconfigured
    makeprogs/installdata.sh
    papsrv/.depend
    papsrv/papsrv.c
    papsrv/papsrv.h
    papsrv/papsrv_authorize.c
    papsrv/papsrv_conf.c
    papsrv/papsrv_query.c
    po/fr-PPRDRV.po
    po/install_mo.sh
    ppad/.depend
    ppad/ppad_conf.c
    ppad/ppad_printer.c
    ppad/ppad_util.c
    ppd/Makefile
    ppd/generic/Makefile
    ppd/generic/generic_master
    ppd/ghostscript/ghostscript_master
    ppd/pagesizes.conf
    ppop/ppop.c
    ppop/ppop_cmds_listq.c
    ppr/.depend
    ppr/ppr.h
    ppr/ppr_dscdoc.c
    ppr/ppr_infile.c
    ppr/ppr_main.c
    ppr/ppr_nest.c
    ppr/ppr_things.c
    pprd/.depend
    pprd/pprd_destid.c
    pprd/pprd_media.c
    pprd/pprd_pprdrv.c
    pprd/pprd_queue.c
    pprdrv/.depend
    pprdrv/pprdrv.c
    pprdrv/pprdrv.h
    pprdrv/pprdrv_custom_hook.c
    pprdrv/pprdrv_fault_debug.c
    pprdrv/pprdrv_feedback.c
    pprdrv/pprdrv_flag.c
    pprdrv/pprdrv_interface.c
    pprdrv/pprdrv_ppd.c
    pprdrv/pprdrv_res.c
    pprdrv/pprdrv_rip.c
    pprdrv/pprdrv_writemon.c
    procsets/Makefile
    procsets/TrinColl-PPR-dmm-nup-3-6
    samba/smb-protos.conf
    uprint/.depend
    vendors/ppd_hp/HP LaserJet 4000 Series
    vendors/ppd_hp/HP LaserJet 4050 Series
    www/index.html
Delete:
    include/cexcept.h
    include/gu_md5.h
    interfaces/README_INTERFACES.txt
    libgu/gu_md5.c
    libgu/gu_wildmat.c
    libscript/interface.sh
    makeprogs/paths.mk
    makeprogs/squeeze.c
    po/PAPSRV.pot
    po/PPR.pot
    po/PPRD.pot
    po/PPRDRV.pot
    po/PPRWWW.pot
    ppd/generic/PPR Generic
    ppd/generic/PPR Generic Level 2
    tests/test-ppad/000-setup.out
    tests/test-ppad/100-interface.out
    tests/test-ppad/200-passthru.out
    tests/test-ppad/999-cleanup.out
    vendors/ppd_hp/HP LaserJet 4050 N
    www/q_icons/00001.pbm
    www/q_icons/00002.pbm
    www/q_icons/00003.pbm
    www/q_icons/Makefile
    www/q_icons/transfig.tex

+++ makeprogs/installdata.sh
@@ -1,6 +1,6 @@
 #
-# mouse:~ppr/src/installdata.sh
-# Copyright 1995--2000, Trinity College Computing Center.
+# mouse:~ppr/src/makeprogs/installdata.sh
+# Copyright 1995--2001, Trinity College Computing Center.
 # Written by David Chappell.
 #
 # Permission to use, copy, modify, and distribute this software and its
@@ -10,7 +10,7 @@
 # documentation.  This software and documentation are provided "as is"
 # without express or implied warranty.
 #
-# Last modified 13 September 2000.
+# Last modified 18 June 2001.
 #
 
 #
@@ -43,11 +43,14 @@ if [ ! -d "$RPM_BUILD_ROOT$DESTDIR" ]
 while [ "$1" != "" ]
     do
     name=`basename "$1"`
-    dest="$DESTDIR/$name"
-    echo "    \"$1\" --> \"$RPM_BUILD_ROOT$dest\""
-    rm -f "$RPM_BUILD_ROOT$dest" || exit 1
-    cp "$1" "$RPM_BUILD_ROOT$dest" || exit 1
-    echo "\"$dest\"" >>`dirname $0`/installed_files_list
+    if [ "$name" != "CVS" ]
+	then
+	dest="$DESTDIR/$name"
+	echo "    \"$1\" --> \"$RPM_BUILD_ROOT$dest\""
+	rm -f "$RPM_BUILD_ROOT$dest" || exit 1
+	cp "$1" "$RPM_BUILD_ROOT$dest" || exit 1
+	echo "\"$dest\"" >>`dirname $0`/installed_files_list
+	fi
     shift
     done
